Move Kubernetes command to seperate file
